backend_map_granted_ring should xs_read page-gref
Xen is inconsistent in its naming of xenstore nodes for the grant
reference.  ring-ref is used for some xen interfaces, but page-gref is
what is used by vkbd.

To let vglass map the vkbd with grants, we need to look up the
xenstore node that actually exists.  Have the code lookup page-gref and
fallback to ring-ref.  vGlass is now the only OpenXT user of this
code, but leave the fallback just in case.

This lets vglass use backend_init_noxc and not open xc_handle or
/dev/xen/privcmd.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>
(cherry picked from commit 28ba8acc028e6dd28d4fc99507ceb0d25990cf41)